requirements.txt
psycopg2-binary>=2.7.4
pytest>=3.4.1
